Add TODOs to cover the minimum functionality.

+++ app/src/main/java/com/blogspot/derefer/fasttracker/Fast.java
@@ -1,16 +1,17 @@
 package com.blogspot.derefer.fasttracker;
 
 public class Fast {
-    private int id;
+    private long id;
     private long begin;
     private long end;
 
-    public Fast(int id, long begin, long end) {
+    public Fast(long id, long begin, long end) {
         this.id = id;
         this.begin = begin;
         this.end = end;
     }
 
+    // TODO: Add string conversion for begin and end
     public String returnFastAsString() {
         return id + ": " + begin + " - " + end;
     }
